+++ importer/utils/csv_parser.py
@@ -1,5 +1,6 @@
 import csv
 from typing import List
+from decimal import Decimal
 
 from models.emissions_model import Sector, Value
 
@@ -13,7 +14,7 @@ def parse_csv_file(file_path: str) -> List[Sector]:
             for key in row:
                 if key.isnumeric():  # check if key is a year
                     year = int(key)
-                    value = float(row[key])
+                    value = Decimal(row[key])
                     values_per_year.append(Value(year=year, value=value))
             sector_data = Sector(country=row['Country'], sector=row['Sector'], parentSector=row['Parent sector'], valuesPerYear=values_per_year)
             sectors.append(sector_data.dict())
